Fixed error

+++ include/matrix.hh
@@ -7,7 +7,7 @@ using matrix = vector<vd>;
 
 matrix transpose(const matrix &m);
 matrix operator*(const matrix &A, const matrix &B);
-matrix operator*(int a, const matrix &A);
+matrix operator*(double a, const matrix &A);
 matrix takeColumns(const vi &take, const matrix &A);
 matrix identityMatrix(int k);
 vd column(const matrix &A, int idx);
